Ugly first Values.
I have to restructure my ExprModel.
Once my Hole is gone, I have no more type Information.
I need to have a "data WithType", or similar, for type
annotations.

Type should be viewed in lambda bodys and in Record
Fields (That part is working fine).

Some constructive TODO: I need to not only define Record TYPES,
but Record VALUES now, too.

Oh, and one more thing: I have to be able to somehow "link"
values to each other, for example link the name of a parameter
to all places it is used (even if it can only be changed in
the lambda abstraction).